[import-w3c-tests] Make TestParser use the WPT SourceFile https://bugs.webkit.org/show_bug.cgi?id=263673
Reviewed by Jonathan Bedard.

Rather than reimplementing all this logic ourselves, we should just use
the upstream WPT code to classify files. This does vendor a new
(partial) copy (from upstream f54724aafa), to pull in the last version
of the WPT tools prior to them dropping Python 2 support.

This vastly reduces the amount of logic webkitpy.w3c.test_parser has in
it, though we maintain our own copy of fuzzy_metadata (now directly in
webkitpy.layout_tests.controllers.single_test_runner), given we need
different semantics when running tests (specifically: we want it to find
fuzzy data for tests WPT doesn't believe are reftests).

* Tools/Scripts/webkitpy/layout_tests/controllers/single_test_runner.py:
(SingleTestRunner._fuzzy_metadata_for_file):
(SingleTestRunner._fuzzy_tolerance_for_reference):
* Tools/Scripts/webkitpy/layout_tests/controllers/single_test_runner_unittest.py:
(SingleTestRunnerTest):
(test_nameless_fuzzy_data):
(test_range_fuzzy_data):
(test_nameless_range_fuzzy_data):
(test_per_ref_fuzzy_data):
* Tools/Scripts/webkitpy/thirdparty/wpt: Added, from upstream f54724aafa.

Canonical link: https://commits.webkit.org/277673@main
