+++ Tools/Scripts/webkitpy/w3c/test_parser_unittest.py
@@ -42,7 +42,7 @@ options = {'all': False, 'no_overwrite': False}
 class TestParserTest(unittest.TestCase):
 
     def test_analyze_test_reftest_one_match(self):
-        test_html = """<head>
+        test_html = b"""<head>
 <link rel="match" href="green-box-ref.xht" />
 </head>
 """
@@ -58,7 +58,7 @@ class TestParserTest(unittest.TestCase):
         self.assertFalse('jstest' in test_info.keys(), 'test should not have been analyzed as a jstest')
 
     def test_analyze_test_reftest_one_mismatch(self):
-        test_html = """<head>
+        test_html = b"""<head>
 <link rel="mismatch" href="green-box-ref.xht" />
 </head>
 """
@@ -74,7 +74,7 @@ class TestParserTest(unittest.TestCase):
         self.assertFalse('jstest' in test_info.keys(), 'test should not have been analyzed as a jstest')
 
     def test_analyze_test_reftest_multiple_matches(self):
-        test_html = """<head>
+        test_html = b"""<head>
 <link rel="match" href="green-box-ref.xht" />
 <link rel="match" href="blue-box-ref.xht" />
 <link rel="match" href="orange-box-ref.xht" />
@@ -98,7 +98,7 @@ class TestParserTest(unittest.TestCase):
         )
 
     def test_analyze_test_reftest_match_and_mismatch(self):
-        test_html = """<head>
+        test_html = b"""<head>
 <link rel="match" href="green-box-ref.xht" />
 <link rel="match" href="blue-box-ref.xht" />
 <link rel="mismatch" href="orange-box-notref.xht" />
@@ -126,12 +126,12 @@ class TestParserTest(unittest.TestCase):
     def test_analyze_test_reftest_with_ref_support_Files(self):
         """ Tests analyze_test() using a reftest that has refers to a reference file outside of the tests directory and the reference file has paths to other support files """
 
-        test_html = """<html>
+        test_html = b"""<html>
 <head>
-<link rel="match" href="../reference/green-box-ref.xht" />
+<link rel="match" href="../../reference/green-box-ref.xht" />
 </head>
 """
-        ref_html = """<head>
+        ref_html = b"""<head>
 <link href="support/css/ref-stylesheet.css" rel="stylesheet" type="text/css">
 <style type="text/css">
     background-image: url("../../support/some-image.png")
@@ -144,7 +144,7 @@ class TestParserTest(unittest.TestCase):
 </html>
 """
         test_path = os.path.join(os.path.sep, 'some', 'madeup', 'path')
-        parser = TestParser(options, os.path.join(test_path, 'somefile.html'))
+        parser = TestParser(options, os.path.join(test_path, 'test', 'somefile.html'), source_root_directory=test_path)
         test_info = parser.analyze_test(test_contents=test_html, ref_contents=ref_html)
 
         self.assertNotEqual(test_info, None, 'did not find a test')
@@ -158,7 +158,7 @@ class TestParserTest(unittest.TestCase):
     def test_analyze_jstest(self):
         """ Tests analyze_test() using a jstest """
 
-        test_html = """<head>
+        test_html = b"""<head>
 <link href="/resources/testharness.css" rel="stylesheet" type="text/css">
 <script src="/resources/testharness.js"></script>
 </head>
@@ -177,7 +177,7 @@ class TestParserTest(unittest.TestCase):
     def test_analyze_manual_wpt_test(self):
         """ Tests analyze_test() using a manual jstest """
 
-        test_html = """<head>
+        test_html = b"""<head>
 <link href="/resources/testharness.css" rel="stylesheet" type="text/css">
 <script src="/resources/testharness.js"></script>
 </head>
@@ -197,17 +197,17 @@ class TestParserTest(unittest.TestCase):
 
         parser = TestParser(options, os.path.join(test_path, 'somefile-ref-manual.html'))
         test_info = parser.analyze_test(test_contents=test_html)
-        self.assertFalse('manualtest' in test_info.keys() and test_info['manualtest'], 'test_info is not manual')
+        self.assertTrue('manualtest' in test_info.keys() and test_info['manualtest'], 'test_info is not manual')
 
         ref_test_path = os.path.join(os.path.sep, 'some', 'madeup', 'path', 'reference')
         parser = TestParser(options, os.path.join(ref_test_path, 'somefile-manual.html'))
         test_info = parser.analyze_test(test_contents=test_html)
-        self.assertFalse('manualtest' in test_info.keys() and test_info['manualtest'], 'test_info is not manual')
+        self.assertTrue('manualtest' in test_info.keys() and test_info['manualtest'], 'test_info is not manual')
 
     def test_analyze_manual_reference_wpt_test(self):
         """ Tests analyze_test() using a manual reference test """
 
-        test_html = """<head>
+        test_html = b"""<head>
 <link rel="match" href="reference/somefile-manual-ref.html">
 <link href="/resources/testharness.css" rel="stylesheet" type="text/css">
 <script src="/resources/testharness.js"></script>
@@ -224,16 +224,16 @@ class TestParserTest(unittest.TestCase):
         test_path = os.path.join(os.path.sep, 'some', 'madeup', 'path')
         parser = TestParser(options, os.path.join(test_path, 'somefile.html'))
 
-        for content in ["", "flag1", "flag1 flag2", "flag1 flag2 flag3", "asis"]:
-            test_html = """
+        for content in [b"", b"flag1", b"flag1 flag2", b"flag1 flag2 flag3", b"asis"]:
+            test_html = b"""
 <head>
   <meta name="flags" content="%s">
 </head>""" % content
             test_info = parser.analyze_test(test_contents=test_html)
             self.assertEqual(test_info, None, 'test_info should be None')
 
-        for flag in ["animated", "font", "history", "interact", "paged", "speech", "userstyle"]:
-            test_html = """
+        for flag in [b"animated", b"font", b"history", b"interact", b"paged", b"speech", b"userstyle"]:
+            test_html = b"""
 <head>
   <meta name="flags" content="flag1 flag2">
   <meta name="flags" content="flag3 %s flag4 flag5">
@@ -244,7 +244,7 @@ class TestParserTest(unittest.TestCase):
             self.assertTrue(test_info['manualtest'], 'test with CSS flag %s should be manual' % flag)
 
     def test_analyze_crash_test(self):
-        test_html = """<body></body>
+        test_html = b"""<body></body>
 """
         test_path = os.path.join(os.path.sep, 'some', 'madeup', 'path')
         parser = TestParser(options, os.path.join(test_path, 'somefile-crash.html'))
@@ -261,7 +261,7 @@ class TestParserTest(unittest.TestCase):
     def test_analyze_pixel_test_all_true(self):
         """ Tests analyze_test() using a test that is neither a reftest or jstest with all=False """
 
-        test_html = """<html>
+        test_html = b"""<html>
 <head>
 <title>CSS Test: DESCRIPTION OF TEST</title>
 <link rel="author" title="NAME_OF_AUTHOR" />
@@ -296,7 +296,7 @@ CONTENT OF TEST
     def test_analyze_pixel_test_all_false(self):
         """ Tests analyze_test() using a test that is neither a reftest or jstest, with -all=False """
 
-        test_html = """<html>
+        test_html = b"""<html>
 <head>
 <title>CSS Test: DESCRIPTION OF TEST</title>
 <link rel="author" title="NAME_OF_AUTHOR" />
@@ -333,10 +333,10 @@ CONTENT OF TEST
     def test_reference_test(self):
         """ Tests analyze_test() using a test that is a reference file having a <link rel="match"> tag"""
 
-        test_html = """<html>
+        test_html = b"""<html>
 <head>
 <title>CSS Test: DESCRIPTION OF TEST</title>
-<link rel="match" href="test-ref.html" />
+<link rel="match" href="test-prime-ref.html" />
 <link rel="author" title="NAME_OF_AUTHOR" />
 <style type="text/css"><![CDATA[
 CSS FOR TEST
@@ -351,80 +351,5 @@ CONTENT OF TEST
         parser = TestParser(options, os.path.join(test_path, 'test-ref.html'))
         test_info = parser.analyze_test(test_contents=test_html)
 
-        self.assertTrue('referencefile' in test_info, 'test should be detected as reference file')
-
-    def _fuzzy_metadata_from_test_with_contents(self, test_contents):
-        test_path = os.path.join(os.path.sep, 'some', 'madeup', 'path')
-        parser = TestParser({'all': True}, os.path.join(test_path, 'test-fuzzy.html'))
-        parser.analyze_test(test_contents=test_contents)  # We need to analyze the test_contents
-        return parser.fuzzy_metadata()
-
-    def test_simple_fuzzy_data(self):
-        """ Tests basic form of fuzzy_metadata()"""
-
-        test_html = """<html><head>
-<meta name=fuzzy content="maxDifference = 15 ; totalPixels = 300">
-</head>
-<body>CONTENT OF TEST</body></html>
-"""
-        actual_fuzzy = self._fuzzy_metadata_from_test_with_contents(test_html)
-
-        expected_fuzzy = {None: [[15, 15], [300, 300]]}
-        self.assertEqual(actual_fuzzy, expected_fuzzy, 'fuzzy data did not match expected')
-
-    def test_nameless_fuzzy_data(self):
-        """ Tests fuzzy_metadata() in short form"""
-
-        test_html = """<html><head>
-<meta name=fuzzy content=" 15 ; 300 ">
-</head>
-<body>CONTENT OF TEST</body></html>
-"""
-        actual_fuzzy = self._fuzzy_metadata_from_test_with_contents(test_html)
-        expected_fuzzy = {None: [[15, 15], [300, 300]]}
-        self.assertEqual(actual_fuzzy, expected_fuzzy, 'fuzzy data did not match expected')
-
-    def test_range_fuzzy_data(self):
-        """ Tests fuzzy_metadata() in range form"""
-
-        test_html = """<html><head>
-<meta name=fuzzy content="maxDifference=5-15;totalPixels =  200 - 300 ">
-</head>
-<body>CONTENT OF TEST</body></html>
-"""
-        actual_fuzzy = self._fuzzy_metadata_from_test_with_contents(test_html)
-
-        expected_fuzzy = {None: [[5, 15], [200, 300]]}
-        self.assertEqual(actual_fuzzy, expected_fuzzy, 'fuzzy data did not match expected')
-
-    def test_nameless_range_fuzzy_data(self):
-        """ Tests fuzzy_metadata() in short range form"""
-
-        test_html = """<html><head>
-<meta name=fuzzy content="5-15;  200 - 300 ">
-</head>
-<body>CONTENT OF TEST</body></html>
-"""
-        actual_fuzzy = self._fuzzy_metadata_from_test_with_contents(test_html)
-
-        expected_fuzzy = {None: [[5, 15], [200, 300]]}
-        self.assertEqual(actual_fuzzy, expected_fuzzy, 'fuzzy data did not match expected')
-
-    def test_per_ref_fuzzy_data(self):
-        """ Tests fuzzy_metadata() with values for difference reference files"""
-
-        test_html = """<html><head>
-<meta name=fuzzy content="5-15;200-300 ">
-<meta name=fuzzy content="close-match-ref.html:5;20">
-<meta name=fuzzy content="worse-match-ref.html: 15;30">
-</head>
-<body>CONTENT OF TEST</body></html>
-"""
-        actual_fuzzy = self._fuzzy_metadata_from_test_with_contents(test_html)
-
-        expected_fuzzy = {
-            None: [[5, 15], [200, 300]],
-            'close-match-ref.html': [[5, 5], [20, 20]],
-            'worse-match-ref.html': [[15, 15], [30, 30]]
-        }
-        self.assertEqual(actual_fuzzy, expected_fuzzy, 'fuzzy data did not match expected')
+        self.assertFalse('referencefile' in test_info, 'test should not be detected as reference file')
+        self.assertTrue('match_reference' in test_info.keys())
